Move `ExecutionCidState` to separate module
And also refactor some funcs.
