+++ air/src/execution_step/execution_context/context.rs
@@ -14,27 +14,17 @@
  * limitations under the License.
  */
 
+use super::ExecutionCidState;
 use super::LastError;
 use super::LastErrorDescriptor;
 use super::Scalars;
 use super::Streams;
-use crate::execution_step::RcSecurityTetraplet;
-use crate::execution_step::ValueAggregate;
-use crate::JValue;
-use crate::UncatchableError;
 
 use air_execution_info_collector::InstructionTracker;
-use air_interpreter_cid::CidCalculationError;
-use air_interpreter_cid::CID;
-use air_interpreter_data::CanonCidAggregate;
 use air_interpreter_data::CidInfo;
-use air_interpreter_data::CidTracker;
 use air_interpreter_data::GlobalStreamGens;
 use air_interpreter_data::RestrictedStreamGens;
-use air_interpreter_data::ServiceResultAggregate;
-use air_interpreter_data::TracePos;
 use air_interpreter_interface::*;
-use polyplets::SecurityTetraplet;
 
 use std::rc::Rc;
 
@@ -144,115 +134,6 @@ pub(crate) struct ExecCtxIngredients {
     pub(crate) last_call_request_id: u32,
     pub(crate) restricted_streams: RestrictedStreamGens,
     pub(crate) cid_info: CidInfo,
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct ExecutionCidState {
-    pub value_tracker: CidTracker<JValue>,
-    pub tetraplet_tracker: CidTracker<SecurityTetraplet>,
-    pub canon_tracker: CidTracker<CanonCidAggregate>,
-    pub service_result_agg_tracker: CidTracker<ServiceResultAggregate>,
-}
-
-impl ExecutionCidState {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn insert_value(
-        &mut self,
-        value: Rc<JValue>,
-        tetraplet: RcSecurityTetraplet,
-        argument_hash: Rc<str>,
-    ) -> Result<Rc<CID<ServiceResultAggregate>>, CidCalculationError> {
-        let value_cid = self.value_tracker.record_value(value)?;
-        let tetraplet_cid = self.tetraplet_tracker.record_value(tetraplet)?;
-
-        let service_result_agg = ServiceResultAggregate {
-            value_cid,
-            argument_hash,
-            tetraplet_cid,
-        };
-
-        self.service_result_agg_tracker.record_value(service_result_agg)
-    }
-
-    fn from_cid_info(prev_cid_info: CidInfo, current_cid_info: CidInfo) -> Self {
-        Self {
-            value_tracker: CidTracker::from_cid_stores(prev_cid_info.value_store, current_cid_info.value_store),
-            tetraplet_tracker: CidTracker::from_cid_stores(
-                prev_cid_info.tetraplet_store,
-                current_cid_info.tetraplet_store,
-            ),
-            canon_tracker: CidTracker::from_cid_stores(prev_cid_info.canon_store, current_cid_info.canon_store),
-            service_result_agg_tracker: CidTracker::from_cid_stores(
-                prev_cid_info.service_result_store,
-                current_cid_info.service_result_store,
-            ),
-        }
-    }
-
-    pub(crate) fn get_value_by_cid(&self, cid: &CID<JValue>) -> Result<Rc<JValue>, UncatchableError> {
-        self.value_tracker
-            .get(cid)
-            .ok_or_else(|| UncatchableError::ValueForCidNotFound("value", cid.clone().into()))
-    }
-
-    pub(crate) fn get_tetraplet_by_cid(
-        &self,
-        cid: &CID<SecurityTetraplet>,
-    ) -> Result<RcSecurityTetraplet, UncatchableError> {
-        self.tetraplet_tracker
-            .get(cid)
-            .ok_or_else(|| UncatchableError::ValueForCidNotFound("tetraplet", cid.clone().into()))
-    }
-
-    pub(crate) fn get_canon_value_by_cid(
-        &self,
-        cid: &CID<CanonCidAggregate>,
-    ) -> Result<ValueAggregate, UncatchableError> {
-        let canon_aggregate = self
-            .canon_tracker
-            .get(cid)
-            .ok_or_else(|| UncatchableError::ValueForCidNotFound("canon aggregate", cid.clone().into()))?;
-        let result = self.get_value_by_cid(&canon_aggregate.value)?;
-        let tetraplet = self.get_tetraplet_by_cid(&canon_aggregate.tetraplet)?;
-
-        let fake_trace_pos = TracePos::default();
-        Ok(ValueAggregate {
-            result,
-            tetraplet,
-            trace_pos: fake_trace_pos,
-        })
-    }
-
-    pub(crate) fn get_service_result_agg_by_cid(
-        &self,
-        cid: &CID<ServiceResultAggregate>,
-    ) -> Result<Rc<ServiceResultAggregate>, UncatchableError> {
-        self.service_result_agg_tracker
-            .get(cid)
-            .ok_or_else(|| UncatchableError::ValueForCidNotFound("service result aggregate", cid.clone().into()))
-    }
-
-    pub(crate) fn resolve_service_value(
-        &self,
-        service_result_agg_cid: &CID<ServiceResultAggregate>,
-    ) -> Result<Rc<JValue>, UncatchableError> {
-        let service_result_aggregate = self.get_service_result_agg_by_cid(service_result_agg_cid)?;
-        self.get_value_by_cid(&service_result_aggregate.value_cid)
-    }
-}
-
-impl From<ExecutionCidState> for CidInfo {
-    fn from(value: ExecutionCidState) -> Self {
-        Self {
-            value_store: value.value_tracker.into(),
-            tetraplet_store: value.tetraplet_tracker.into(),
-            canon_store: value.canon_tracker.into(),
-            service_result_store: value.service_result_agg_tracker.into(),
-        }
-    }
 }
 
 use serde::Deserialize;
