+++ air/src/execution_step/execution_context/mod.rs
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+mod cid_state;
 mod context;
 mod last_error;
 mod scalar_variables;
@@ -21,7 +22,7 @@ mod streams_variables;
 
 pub use last_error::*;
 
-pub use context::ExecutionCidState;
+pub use cid_state::ExecutionCidState;
 pub(crate) use context::*;
 pub(crate) use scalar_variables::*;
 pub(crate) use streams_variables::*;
